requirements.txt
anthropic>=0.30
python-dotenv>=1.0
